memperbaiki review dari ka tama, untuk data dummy dari artikel dijadikan menjadi object memperbaiki halaman detail artikel pada recyclerview yang paling bawah sendiri memperbaiki navigate pada halaman artikel agar bisa berfungsi dengan baik membantu task dari mas haryadi yaitu membikin recyclerview dengan 4 data item secara horizontal halaman artikel sudah berfungsi dengan data dummy sementara (- yang belum mungkin filtering tab nya)
