+++ app/src/main/java/com/example/finalproject_chilicare/adapter/PopularCityAdapter.kt
@@ -6,43 +6,27 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.example.finalproject_chilicare.R
-import com.example.finalproject_chilicare.data.models.CurrentWeatherDetails
+import com.example.finalproject_chilicare.data.response.PopularCityResponse
 
-class PopularCityAdapter (private val listCity : List<CurrentWeatherDetails>) : RecyclerView.Adapter<PopularCityAdapter.PopularCityHolder> (){
+class PopularCityAdapter(private val dataList: ArrayList<PopularCityResponse>) : RecyclerView.Adapter<PopularCityAdapter.CityViewHolder>() {
 
-    var onItemClick : ((CurrentWeatherDetails) -> Unit)? = null
-
-    inner class PopularCityHolder (private val view: View) : RecyclerView.ViewHolder(view) {
-
-        fun bindView (popularCity : CurrentWeatherDetails) {
-
-            val nameCity = view.findViewById<TextView>(R.id.textCity)
-
-            nameCity.text = popularCity.city
-
-        }
-
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CityViewHolder {
+        val itemView = LayoutInflater.from(parent.context).inflate(R.layout.card_popular_city, parent, false)
+        return CityViewHolder(itemView)
     }
 
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): PopularCityHolder {
-        val view = LayoutInflater.from(parent.context)
-            .inflate(R.layout.card_popular_city, parent, false)
-        return PopularCityHolder(view)
+    override fun onBindViewHolder(holder: CityViewHolder, position: Int) {
+        val currentItem = dataList[position]
+
+        holder.rvTitle.text = currentItem.dataTitle
     }
 
     override fun getItemCount(): Int {
-        return listCity.size
+        return dataList.size
     }
 
-    override fun onBindViewHolder(holder: PopularCityHolder, position: Int) {
-
-        val itemCity = listCity[position]
-        holder.bindView(listCity[position])
-
-        holder.itemView.setOnClickListener {
-            onItemClick?.invoke(itemCity)
-        }
-
-
+    class CityViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
+        val rvTitle: TextView = itemView.findViewById(R.id.titleCity)
     }
+
 }
