+++ app/src/main/java/com/example/finalproject_chilicare/data/response/PopularCityResponse.kt
@@ -1,0 +1,3 @@
+package com.example.finalproject_chilicare.data.response
+
+data class PopularCityResponse(var dataTitle:String)
