+++ app/src/main/java/com/example/finalproject_chilicare/ui/home/DetailArticleActivity.kt
@@ -12,10 +12,8 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.finalproject_chilicare.R
 import com.example.finalproject_chilicare.adapter.CardAdapter
-import com.example.finalproject_chilicare.adapter.TabAdapter
 import com.example.finalproject_chilicare.data.response.CardResponse
-import com.example.finalproject_chilicare.data.response.TabResponse
-
+import com.example.finalproject_chilicare.utils.DummyDataArtikel
 
 class DetailArticleActivity : AppCompatActivity() {
 
@@ -24,19 +22,6 @@ class DetailArticleActivity : AppCompatActivity() {
 
     private lateinit var cardResponses: ArrayList<CardResponse>
 
-    lateinit var cardTitleList: Array<String>
-    lateinit var cardSubtitleList: Array<String>
-    lateinit var cardDescriptionList: Array<String>
-    lateinit var cardDurasibacaList: Array<String>
-    lateinit var cardImageList: Array<Int>
-
-    lateinit var dataImageDetails: Array<Int>
-    lateinit var dataTanggalDetails: Array<String>
-    lateinit var dataWaktuDetails: Array<String>
-    lateinit var dataTitleDetails: Array<String>
-    lateinit var dataSubtitleDetails: Array<String>
-    lateinit var dataWebViewDetails: Array<String>
-
     private lateinit var tvWebView: WebView
     private lateinit var ivKembali: ImageView
 
@@ -44,28 +29,41 @@ class DetailArticleActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_detail_article)
 
-
+        // Menginisialisasi view
         ivKembali = findViewById(R.id.ivKembali)
+        tvWebView = findViewById(R.id.tvWebView)
+        rvCardArticle = findViewById(R.id.rv_cardArticle2)
+
+        // Set listener untuk tombol kembali
         ivKembali.setOnClickListener {
             Intent(this, ArticleActivity::class.java).also {
                 startActivity(it)
             }
         }
 
+        // Mendapatkan data dari intent
         val getData = intent.getParcelableExtra<CardResponse>("android")
         if (getData != null) {
+            // Menetapkan data ke view
             val dataImageDetails: ImageView = findViewById(R.id.ivArticle)
             val dataTanggalDetails: TextView = findViewById(R.id.tvTanggal)
             val dataWaktuDetails: TextView = findViewById(R.id.tvWaktu)
             val dataTitleDetails: TextView = findViewById(R.id.tvJudulartikel)
             val dataSubtitleDetails: TextView = findViewById(R.id.tvTab)
             val dataWebViewDetails: WebView = findViewById(R.id.tvWebView)
+
+            // Menetapkan data ke view
             dataImageDetails.setImageResource(getData.dataImageDetails)
             dataTanggalDetails.text = getData.dataTanggalDetails
             dataWaktuDetails.text = getData.dataWaktuDetails
             dataTitleDetails.text = getData.dataTitleDetails
             dataSubtitleDetails.text = getData.dataSubtitleDetails
             dataWebViewDetails.loadData(getData.dataWebViewDetails, "text/html", "utf-8")
+
+            // Memuat data HTML ke WebView
+            tvWebView.settings.javaScriptEnabled = true
+            tvWebView.webViewClient = WebViewClient()
+            tvWebView.loadDataWithBaseURL(null, getData.dataWebViewDetails, "text/html", "UTF-8", null)
         }
 
         tvWebView = findViewById(R.id.tvWebView)
@@ -98,159 +96,24 @@ class DetailArticleActivity : AppCompatActivity() {
         tvWebView.loadDataWithBaseURL(null, htmlText, "text/html", "UTF-8", null)
 
 
-        cardTitleList = arrayOf(
-            "Menanam",
-            "Bibit",
-            "Hama",
-            "Pestisida",
-            "Menanam",
-            "Menanam",
-            "Petisida",
-            "Petisida",
-            "Petisida",
-            "Petisida"
-        )
-        cardSubtitleList = arrayOf(
-            "Mudah Menanam Cabe Agar Cepat Berbuah",
-            "Tips Memilih Bibit Cabai Yang Berkualitas",
-            "Cegah Hama Pada Tanaman Cabai Anda!",
-            "Membuat Pestisida Alami Tanaman Cabai",
-            "Cara Mudah Menanam Cabai di Pot",
-            "Teknik Budidaya Cabai Rawit",
-            "Teknik Budidaya Cabai Rawit",
-            "Membuat pestisida alami tanaman cabai",
-            "Membuat pestisida alami tanaman cabai",
-            "Membuat pestisida alami tanaman cabai"
-        )
-        cardDescriptionList = arrayOf(
-            "Semakin banyak kegiatan seru di tengah kesibukan, salah satunya bertani",
-            "Bibit yang berkualitas memiliki beberapa ciri-ciri yang bisa chili-cares amati",
-            "Hama dapat menyebabkan gagal panen pada tanaman cabai Anda.",
-            "Meningkatnya kesadaran masyarakat akan efek samping menggunakan pestisida kimiawi",
-            "Meningkatnya kesadaran masyarakat akan efek samping menggunakan pestisida kimiawi",
-            "Cabai rawit atau cabai kecil (Capsicum frutescens) termasuk dalamfamili Solanaceae dan merupakan tanaman berumur panjang (menahun)",
-            "Meningkatnya kesadaran masyarakat akan efek samping menggunakan pestisida kimiawi",
-            "Meningkatnya kesadaran masyarakat akan efek samping menggunakan pestisida kimiawi",
-            "Meningkatnya kesadaran masyarakat akan efek samping menggunakan pestisida kimiawi",
-            "Meningkatnya kesadaran masyarakat akan efek samping menggunakan pestisida kimiawi"
-        )
-        cardDurasibacaList =
-            arrayOf(
-                "3 menit baca",
-                "5 menit baca",
-                "4 menit baca",
-                "5 menit baca",
-                "5 menit baca",
-                "10 menit dibaca",
-                "10 menit dibaca",
-                "10 menit dibaca",
-                "10 menit dibaca",
-                "10 menit dibaca"
-            )
-        cardImageList = arrayOf(
-            R.drawable.gambar_1,
-            R.drawable.gambar_2,
-            R.drawable.gambar_3,
-            R.drawable.gambar_4,
-            R.drawable.gambar_5,
-            R.drawable.gambar_6,
-            R.drawable.gambar_7,
-            R.drawable.gambar_8,
-            R.drawable.gambar_9,
-            R.drawable.gambar_10
-        )
-        dataImageDetails = arrayOf(
-            R.drawable.gambar_1,
-            R.drawable.gambar_2,
-            R.drawable.gambar_3,
-            R.drawable.gambar_4,
-            R.drawable.gambar_5,
-            R.drawable.gambar_6,
-            R.drawable.gambar_7,
-            R.drawable.gambar_8,
-            R.drawable.gambar_9,
-            R.drawable.gambar_10
-        )
-        dataTanggalDetails = arrayOf(
-            "19 Januari 2021",
-            "19 Januari 2021",
-            "19 Januari 2021",
-            "19 Januari 2021",
-            "19 Januari 2021",
-            "19 Januari 2021",
-            "19 Januari 2021",
-            "19 Januari 2021",
-            "19 Januari 2021",
-            "19 Januari 2021"
-        )
-        dataWaktuDetails = arrayOf(
-            "19:00 WIB",
-            "19:00 WIB",
-            "19:00 WIB",
-            "19:00 WIB",
-            "19:00 WIB",
-            "19:00 WIB",
-            "19:00 WIB",
-            "19:00 WIB"
-        )
-        dataTitleDetails = arrayOf(
-            "Mudah Menanam Cabe Agar Cepat Berbuah",
-            "Tips Memilih Bibit Cabai Yang Berkualitas",
-            "Cegah Hama Pada Tanaman Cabai",
-            "Membuat Pestisida Alami Tanaman Cabai",
-            "Cara Mudah Menanam Cabai di Pot",
-            "Teknik Budidaya Cabai Rawit",
-            "Teknik Budidaya Cabai Rawit",
-            "Membuat pestisida alami tanaman cabai"
-        )
-        dataSubtitleDetails = arrayOf(
-            "Menanam",
-            "Bibit",
-            "Hama",
-            "Pestisida",
-            "Menanam",
-            "Menanam",
-            "Petisida",
-            "Petisida",
-            "Petisida",
-            "Petisida"
-        )
-        dataWebViewDetails = Array(cardImageList.size) { "" }
 
-        rvCardArticle = findViewById(R.id.rv_cardArticle2)
+        val cardResponses: MutableList<CardResponse> = DummyDataArtikel.getDummyCardResponses().toMutableList()
+
+        val filteredCardResponses = cardResponses.take(2).toMutableList()
+
+
+        // Inisialisasi dan atur adapter CardAdapter untuk RecyclerView card
+        cardAdapter = CardAdapter(filteredCardResponses)
         rvCardArticle.layoutManager = LinearLayoutManager(this)
         rvCardArticle.setHasFixedSize(true)
-
-        cardResponses = arrayListOf<CardResponse>()
-        getDataCard()
-
-        cardAdapter = CardAdapter(cardResponses)
         rvCardArticle.adapter = cardAdapter
+
+        // Menangani klik item di RecyclerView card
         cardAdapter.onItemClick = {
             Log.d("ArticleActivity", "Clicked item: $it")
             val intent = Intent(this, DetailArticleActivity::class.java)
             intent.putExtra("android", it)
             startActivity(intent)
-        }
-    }
-
-    private fun getDataCard() {
-        cardResponses.clear()
-        for (i in cardTitleList.indices) {
-            val cardResponse = CardResponse(
-                cardTitleList[i],
-                cardSubtitleList[i],
-                cardDescriptionList[i],
-                cardDurasibacaList[i],
-                cardImageList[i],
-                dataImageDetails[i],
-                dataTanggalDetails[i],
-                dataWaktuDetails[i],
-                dataTitleDetails[i],
-                dataSubtitleDetails[i],
-                dataWebViewDetails[i]
-            )
-            cardResponses.add(cardResponse)
         }
     }
 }
