+++ app/src/main/res/layout/activity_detail_article.xml
@@ -42,6 +42,7 @@
         android:layout_weight="1"
         android:layout_marginTop="20dp"
         android:layout_marginBottom="20dp"
+        android:paddingBottom="100dp"
         app:layout_constraintTop_toBottomOf="@+id/ivKembali"
         >
