+++ app/src/main/res/layout/activity_search_popular_city.xml
@@ -4,17 +4,20 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.home.SearchPopularCityActivity">
+    tools:context=".ui.home.SearchPopularCityActivity"
+    android:paddingHorizontal="16dp"
+    android:background="@color/BackgroundHalaman"
+    >
 
     <ImageView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginLeft="16dp"
         android:layout_marginTop="16dp"
         android:id="@+id/iconBack"
         android:src="@drawable/baseline_arrow_back_24"
         app:layout_constraintLeft_toLeftOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+        app:layout_constraintTop_toTopOf="parent"
+        />
 
     <TextView
         android:id="@+id/textKelolaKota"
@@ -22,19 +25,19 @@
         android:layout_height="wrap_content"
         android:text="Cari Kota"
         app:layout_constraintLeft_toRightOf="@+id/iconBack"
-        app:layout_constraintTop_toTopOf="parent"
-        android:layout_marginLeft="16dp"
-        android:layout_marginTop="16dp"
+        app:layout_constraintTop_toTopOf="@+id/iconBack"
+        app:layout_constraintBottom_toBottomOf="@+id/iconBack"
+        android:layout_marginLeft="24dp"
         android:textSize="16sp"
-        android:textStyle="bold"
-        android:textColor="@color/black"/>
+        android:textColor="@color/black"
+        android:fontFamily="@font/plusjakartasans_bold_700"
+        />
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/tilCariArtikel"
-        android:layout_width="350dp"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="18dp"
-        android:layout_marginLeft="28dp"
+        android:layout_marginTop="24dp"
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintTop_toBottomOf="@id/textKelolaKota"
         app:hintEnabled="false"
@@ -63,18 +66,19 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/tilCariArtikel"
         android:text="Kota Popular"
-        android:layout_marginTop="8dp"
+        android:textSize="16sp"
+        android:fontFamily="@font/plusjakartasans_bold_700"
+        android:layout_marginTop="24dp"
         android:textColor="@color/black"
-        android:textStyle="bold"
-        android:layout_marginLeft="12dp"/>
+        />
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rvPopularCity"
-        tools:listitem="@layout/card_popular_city"
-        android:layout_width="wrap_content"
+        android:id="@+id/rv_CariKota"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         app:layout_constraintTop_toBottomOf="@id/textKotaPopular"
         app:layout_constraintLeft_toLeftOf="parent"
-        android:layout_marginTop="12dp"/>
+        android:layout_marginTop="16dp"
+        />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
